[RELEASE] Youthful Yttrofluorite

+++ libs/common/src/lib/events/entity-status-update.event-processor.ts
@@ -22,7 +22,10 @@ import {
   Form,
   FormQuestion,
   FormSubmission,
+  NurseryReport,
   ProjectPitch,
+  ProjectReport,
+  SiteReport,
   Task,
   UpdateRequest
 } from "@terramatch-microservices/database/entities";
@@ -39,7 +42,6 @@ import {
   STARTED,
   STATUS_DISPLAY_STRINGS
 } from "@terramatch-microservices/database/constants/status";
-import { InternalServerErrorException } from "@nestjs/common";
 import { LARAVEL_MODELS } from "@terramatch-microservices/database/constants/laravel-types";
 import { Model } from "sequelize-typescript";
 import { getLinkedFieldConfig } from "../linkedFields";
@@ -268,7 +270,10 @@ export class EntityStatusUpdate extends EventProcessor {
   private async updateActions() {
     this.logger.log(`Updating actions [${JSON.stringify({ model: this.model.constructor.name, id: this.model.id })}]`);
     const entity = this.model as EntityModel;
-    await Action.for(entity).destroy({ where: { type: "notification" } });
+
+    if (await this.canActionBeDeleted()) {
+      await Action.for(entity).destroy({ where: { type: "notification" } });
+    }
 
     if (entity.status !== AWAITING_APPROVAL) {
       const action = new Action();
@@ -335,8 +340,7 @@ export class EntityStatusUpdate extends EventProcessor {
       return;
     }
 
-    const modelWithTaskId = model as ReportModel & { taskId: number | null };
-    const { taskId } = modelWithTaskId;
+    const { taskId } = model as { taskId: number | null };
     if (taskId == null) {
       this.logger.warn(`No task found for status changed report [${model.constructor.name}, ${model.id}]`);
       return;
@@ -348,7 +352,8 @@ export class EntityStatusUpdate extends EventProcessor {
       include: [
         { association: "projectReport", attributes },
         { association: "siteReports", attributes },
-        { association: "nurseryReports", attributes }
+        { association: "nurseryReports", attributes },
+        { association: "srpReports", attributes }
       ]
     });
     if (task == null) {
@@ -356,14 +361,12 @@ export class EntityStatusUpdate extends EventProcessor {
       return;
     }
 
-    if (task.status === DUE) {
-      // No further processing needed; nothing automatic happens until the task has been submitted.
-      return;
-    }
-
-    const reports = flatten<ReportModel | null>([task.projectReport, task.siteReports, task.nurseryReports]).filter(
-      report => report != null
-    );
+    const reports = flatten<ReportModel | null>([
+      task.projectReport,
+      task.siteReports,
+      task.nurseryReports,
+      task.srpReports
+    ]).filter(isNotNull);
 
     const reportStatuses = uniq(reports.map(({ status }) => status));
     if (reportStatuses.length === 1 && reportStatuses[0] === APPROVED) {
@@ -372,7 +375,7 @@ export class EntityStatusUpdate extends EventProcessor {
     }
 
     if (reportStatuses.includes(DUE) || reportStatuses.includes(STARTED)) {
-      throw new InternalServerErrorException(`Task has unsubmitted reports [${taskId}]`);
+      return; // NOOP
     }
 
     const moreInfoReport = reports.find(
@@ -386,8 +389,49 @@ export class EntityStatusUpdate extends EventProcessor {
       return;
     }
 
-    // If there are no reports or update requests in needs-more-information, the only option left is that
-    // something is in awaiting-approval.
+    // At this point, there are no reports in due, started or needs-more-information, but they're
+    // not all approved, so at least one report is in awaiting-approval.
     await task.update({ status: AWAITING_APPROVAL });
+  }
+
+  private async canActionBeDeleted() {
+    if (
+      this.model instanceof ProjectReport ||
+      this.model instanceof SiteReport ||
+      this.model instanceof NurseryReport
+    ) {
+      if (this.model.taskId == null) {
+        this.logger.error(`No taskId found for report [${this.model.constructor.name}, ${this.model.id}]`);
+        return false;
+      }
+      const task = await Task.findOne({
+        where: { id: this.model.taskId },
+        include: [
+          {
+            association: "projectReport",
+            attributes: ["id", "status"],
+            where: { status: { [Op.notIn]: [DUE, NEEDS_MORE_INFORMATION] } }
+          },
+          {
+            association: "siteReports",
+            attributes: ["id", "status"],
+            where: { status: { [Op.notIn]: [DUE, NEEDS_MORE_INFORMATION] } }
+          },
+          {
+            association: "nurseryReports",
+            attributes: ["id", "status"],
+            where: { status: { [Op.notIn]: [DUE, NEEDS_MORE_INFORMATION] } }
+          }
+        ]
+      });
+      if (task == null) {
+        this.logger.error(`No task found for report [${this.model.constructor.name}, ${this.model.id}]`);
+        return false;
+      }
+      if (task.projectReport == null && task.siteReports?.length === 0 && task.nurseryReports?.length === 0) {
+        return false;
+      }
+    }
+    return true;
   }
 }
